HomeServer/requirements.txt
asgiref==3.3.2
bcrypt==3.1.7
certifi==2019.11.28
chardet==3.0.4
click==8.0
cryptography==2.8
Django==3.2
gunicorn==20.1.0
keyring==18.0.1
MarkupSafe==1.1.0
oauthlib==3.1.0
pexpect==4.6.0
Pillow==7.0.0
pycairo==1.16.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
python-dateutil==2.7.3
pytz==2019.3
PyYAML==5.3.1
requests==2.22.0
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.1
urllib3==1.25.8
uWSGI==2.0.19.1
python-kasa==0.4.0
whitenoise==6.2.0
psycopg2-binary==2.9.5
astral==3.2
tapo==0.8.1
